Rubik's cube solver written in Python
Rubik's cube solver written in Python for 15-112 programming class at
Carnegie Mellon that takes input from the user by allowing the user to
"paint" each side of their scrambled cube, and then provides step by
step instructions on how to solve the cube. The cube configuration is
updated in real time, giving the user the ability to make sure that
their physical cube is matching the cube in the program after each move
is done.